readme
This problem is an extension of the standard bounded-buffer problem, which is a traditional synchronization problem.  Typically, we have one producer process and one consumer process, which share a buffer.  The producer writes items into the buffer and the consumer reads items back out.   In a multi-threaded environment, we cannot assume the order in which the producer and consumer run, and since they share a data structure and possibly more variables, we must use synchronization mechanisms in order to avoid race conditions.  Here is some example code that we might use to solve this problem:  Note that in points to the next free spot in the buffer, and out points to the first filled spot. in is initialized to 0, and out to 0.